login-logout page

+++ EmployeeManagement.Web/Pages/DisplayEmployeeBase.cs
@@ -2,6 +2,7 @@
 using EmployeeManagement.Web.Services;
 using Microsoft.AspNetCore.Components;
 
+
 namespace EmployeeManagement.Web.Pages
 {
     public class DisplayEmployeeBase : ComponentBase
@@ -24,11 +25,20 @@ namespace EmployeeManagement.Web.Pages
         [Inject]
         public NavigationManager NavigationManager { get; set; }
 
-        protected async Task Delete_Click()
+        protected PragimTech.Components.ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
+            DeleteConfirmation.Show();
+
+        }
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
             await EmployeeService.DeleteEmployee(Employee.EmployeeId);
             await OnEmployeeDeleted.InvokeAsync(Employee.EmployeeId);
-            //NavigationManager.NavigateTo("/", true);
+            }
         }
         protected async Task CheckBoxChanged(ChangeEventArgs e)
         {
